Add poison interface to graphql connections
